navbar login/logout buttons fixed

+++ client/src/components/Navbar/Navbar.js
@@ -1,25 +1,29 @@
 import React, { useState, useEffect } from "react";
-import { Link, useNavigate, useLocation } from "react-router-dom";
+import { Link, useLocation } from "react-router-dom";
 import { FaUser } from "react-icons/fa";
 import "./Navbar.css";
 import Searchbar from "../Searchbar/Searchbar";
 import logo from "../../assets/logo.png";
 import axios from "axios";
 
-const Navbar = ({loggedIn, setLoggedIn}) => {
+const Navbar = () => {
   const location = useLocation();
   const [isActive, setisActive] = useState(false);
   const [toggle, setToggle] = useState(false);
   const [pathname, setPathname] = useState(location.pathname);
-  const navigate = useNavigate();
+  const [loggedIn, setLoggedIn] = useState(
+    localStorage.accessToken ? true : false
+  );
 
   useEffect(() => {
     setPathname(location.pathname);
   }, [location]);
 
   useEffect(() => {
-    setLoggedIn(loggedIn)
-  }, [loggedIn])
+    setLoggedIn(localStorage.accessToken ? true : false);
+    console.log(localStorage);
+    console.log(loggedIn);
+  }, [localStorage.accessToken]);
 
   const handler = () => {
     setToggle((current) => !current);
@@ -29,15 +33,20 @@ const Navbar = ({loggedIn, setLoggedIn}) => {
   const logout = async () => {
     try {
       await axios.delete("/logout");
-      setLoggedIn(false);
-      navigate("/account")
-      localStorage.clear();
     } catch (err) {
       console.log(err);
     }
   };
 
-  // useEffect(() => {}, [toggle]);
+  const loginHandler = () => {
+    if (localStorage.accessToken) {
+      setLoggedIn(true);
+      logout();
+      localStorage.clear();
+    } else {
+      setLoggedIn(false);
+    }
+  };
 
   return (
     <nav className="navbar">
@@ -49,24 +58,23 @@ const Navbar = ({loggedIn, setLoggedIn}) => {
       <div className="navigation">
         <Searchbar />
         <ul className="nav-links">
-          {["filter search", "roulette", "analytics", "account"].map((item) => {
-            return (
-              <li
-                className={`link ${
-                  pathname === `/${item.replace(" ", "-")}` ? "active" : ""
-                }`}
-                key={`${item}`}
-              >
-                {item === "account" && loggedIn ? (
-                  <a className="link" onClick={logout}>log out</a>
-                ) : (
-                  <Link to={`/${item.replace(" ", "-")}`}>
-                    {item === "account" && !loggedIn ? <FaUser /> : item}
-                  </Link>
-                )}
-              </li>
-            );
-          })}
+          {["filter search", "roulette", "analytics"].map((item) => (
+            <li
+              className={`link ${pathname === `/${item.replace(" ", "-")}` ? "active" : ""}`}
+              key={`${item}`}
+            >
+              <Link to={`/${item.replace(" ", "-")}`}>{item}</Link>
+            </li>
+          ))}
+          <li className={`link ${pathname === "/account" ? "active" : ""}`}>
+            {loggedIn ? (
+              <a className="link" href="/account" onClick={loginHandler}>
+                log out
+              </a>
+            ) : (
+              <a href="/account" onClick={loginHandler}><FaUser /></a>
+            )}
+          </li>
         </ul>
 
         <div className="hambrgr-menu flex-center" onClick={handler}>
@@ -92,8 +100,8 @@ const Navbar = ({loggedIn, setLoggedIn}) => {
 
         {toggle && (
           <div className="nav-menu flex-center">
-            <ul>
-              {["filter search", "roulette", "analytics", "account"].map(
+          <ul>
+              {["filter search", "roulette", "analytics"].map(
                 (item) => (
                   <li
                     className={`nav-menu-links ${
@@ -107,6 +115,17 @@ const Navbar = ({loggedIn, setLoggedIn}) => {
                   </li>
                 )
               )}
+              <li className={`nav-menu-links`}>
+                {loggedIn ? (
+                  <Link onClick={() => {loginHandler(); handler()}} to="/account">
+                    Log Out
+                  </Link>
+                ) : (
+                  <Link onClick={() => {loginHandler(); handler()}} to="/account">
+                    Log In / Sign Up
+                  </Link>
+                )}
+              </li>
             </ul>
           </div>
         )}
